moving the function to payload/utils
