+++ apps/charterafrica/src/payload/utils/mapLinkAndNameforContributorsInToolsCollection.js
@@ -1,4 +1,4 @@
-import getPageUrlUsingPayload from "../../lib/data/common/getPageUrlUsingPayload";
+import getPageUrlUsingPayload from "./getPageUrlUsingPayload";
 
 const mapLinkAndNameforContributorsInToolsCollection = async ({
   doc,
