+++ apps/charterafrica/src/payload/utils/mapLinkToHrefForOrgCollection.js
@@ -1,4 +1,4 @@
-import getPageUrlUsingPayload from "../../lib/data/common/getPageUrlUsingPayload";
+import getPageUrlUsingPayload from "./getPageUrlUsingPayload";
 
 const mapLinkToHrefForOrgCollection = async ({ doc, req: { payload } }) => {
   try {
